fix warnings

+++ 2-rgb_to_gray/src/rgb_to_gray_kernel.cu
@@ -27,8 +27,8 @@ namespace cuda {
 
 void rgb_to_gray(const std::uint8_t* src, std::uint8_t* dst,
                  const std::size_t width, const std::size_t height) {
-    const auto grid_dim = dim3{height};
-    const auto block_dim = dim3{width};
+    const auto grid_dim = dim3{static_cast<std::uint32_t>(height)};
+    const auto block_dim = dim3{static_cast<std::uint32_t>(width)};
 
     rgb_to_gray_kernel<<<grid_dim, block_dim>>>(src, dst, width, height);
     CUDASafeCall();
