+++ 5-rgb_to_hsv/src/rgb_to_hsv_kernel.cu
@@ -45,8 +45,8 @@ namespace cuda {
 
 void rgb_to_hsv(const std::uint8_t* const src, std::uint8_t* const dst,
                 const std::size_t width, const std::size_t height) {
-    const auto block_dim = dim3{height};
-    const auto thread_dim = dim3{width};
+    const auto block_dim = dim3{static_cast<std::uint32_t>(height)};
+    const auto thread_dim = dim3{static_cast<std::uint32_t>(width)};
 
     rgb_to_hsv_kernel<<<block_dim, thread_dim>>>(src, dst, width, height);
     CUDASafeCall();
