+++ 6-color_reduction/src/color_reduction_kernel.cu
@@ -26,8 +26,8 @@ namespace cuda {
 
 void color_reduction(const std::uint8_t* const src, std::uint8_t* const dst,
                      const std::size_t width, const std::size_t height) {
-    const auto block_dim = dim3{height};
-    const auto thread_dim = dim3{width};
+    const auto block_dim = dim3{static_cast<std::uint32_t>(height)};
+    const auto thread_dim = dim3{static_cast<std::uint32_t>(width)};
 
     color_reduction_kernel<<<block_dim, thread_dim>>>(src, dst, width, height);
     CUDASafeCall();
